+++ RELEASE_NOTES.md
@@ -1,5 +1,25 @@
 # Changelog
 
-## \[10.13.2\] - 31.03.2025
+## \[10.14.0\] - 03.04.2025
 
-Hot fix for the Atlas CDN being down. We're now installing atlas cions locally, and modifying the way it's imported.
+## Components
+
+- Alerts
+  - UI update
+  - Success alert icon update breaking change: the icon name has changed, now using `class="swepay-icon-check-circle-filled-customizable"`
+  - new alert variant (!) : Feedback variant
+
+## Identity
+
+- Flags
+  - split of flags in 2 lists: a short-list, automatically imported via CSS, ready to use simply by using the CSS class, as it was before. about 10 countries flags. For both ratios square + 4/3
+  - list of flags in short-list : ["dk", "ee", "eu", "fi", "gb", "is", "lt", "lv", "no", "se", "us"]
+  - an extended list (the other 240 countries usually never used by any projects in here) that can be imported via your own custom CSS. Examples are added to the documentation.
+
+## PayEx side UI changes
+
+## Technical changes
+
+## Chores
+
+- updated dependencies